Extract fetchTerminatorWebcams into app/lib
Move the SQL helper out of the API route file so non-route callers
(e.g. the windy cron handler) don't import from a route module.
